Move the incremental code discovery support to a new module
This is all pretty self-contained and doesn't need to complicate the
InlineCallee code
