+++ pate.cabal
@@ -77,6 +77,7 @@ library
                        Pate.Config,
                        Pate.Discovery,
                        Pate.Verification,
+                       Pate.Verification.DemandDiscovery,
                        Pate.Verification.Domain,
                        Pate.Verification.ExternalCall,
                        Pate.Verification.InlineCallee,
